refactor form tests to test in loginless rather than input components
